Move & mark functions as private
