+++ tests/unit/cmk/gui/graphing/test_rrd_fetch.py
@@ -3,7 +3,7 @@
 # This file is part of Checkmk (https://checkmk.com). It is subject to the terms and
 # conditions defined in the file COPYING, which is part of this source code package.
 
-from collections.abc import Iterator
+from collections.abc import Iterable, Iterator, Mapping
 from contextlib import contextmanager
 
 import pytest
@@ -22,7 +22,12 @@ from cmk.gui.graphing._graph_specification import (
     MetricOpRRDSource,
 )
 from cmk.gui.graphing._graph_templates import TemplateGraphSpecification
-from cmk.gui.graphing._rrd_fetch import fetch_rrd_data_for_graph, translate_and_merge_rrd_columns
+from cmk.gui.graphing._rrd_fetch import (
+    _reverse_translate_into_all_potentially_relevant_metrics,
+    fetch_rrd_data_for_graph,
+    translate_and_merge_rrd_columns,
+)
+from cmk.gui.graphing._utils import CheckMetricEntry
 from cmk.gui.time_series import TimeSeries, TimeSeriesValues
 from cmk.gui.utils.temperate_unit import TemperatureUnit
 
@@ -265,4 +270,97 @@ def test_translate_and_merge_rrd_columns_unit_conversion(
     ) == TimeSeries(
         expected_data_points,
         time_window=(1682324400, 1682497800, 600),
+    )
+
+
+@pytest.mark.parametrize(
+    ["canonical_name", "current_version", "all_translations", "expected_result"],
+    [
+        pytest.param(
+            MetricName("my_metric"),
+            123,
+            [
+                {
+                    MetricName("some_metric_1"): {"scale": 10},
+                    MetricName("some_metric_2"): {
+                        "scale": 10,
+                        "name": MetricName("new_metric_name"),
+                    },
+                }
+            ],
+            {MetricName("my_metric")},
+            id="no applicable translations",
+        ),
+        pytest.param(
+            MetricName("my_metric"),
+            2030020100,
+            [
+                {
+                    MetricName("some_metric_1"): {"scale": 10},
+                    MetricName("old_name_1"): {
+                        "scale": 10,
+                        "name": MetricName("my_metric"),
+                    },
+                },
+                {
+                    MetricName("old_name_1"): {
+                        "name": MetricName("my_metric"),
+                    },
+                },
+                {
+                    MetricName("old_name_2"): {
+                        "name": MetricName("my_metric"),
+                    },
+                    MetricName("irrelevant"): {"name": MetricName("still_irrelevant")},
+                },
+                {
+                    MetricName("old_name_deprecated"): {
+                        "name": MetricName("my_metric"),
+                        "deprecated": "2.0.0i1",
+                    },
+                },
+            ],
+            {
+                MetricName("my_metric"),
+                MetricName("old_name_1"),
+                MetricName("old_name_2"),
+            },
+            id="some applicable and one deprecated translation",
+        ),
+        pytest.param(
+            MetricName("my_metric"),
+            2030020100,
+            [
+                {
+                    MetricName("old_name_1"): {
+                        "name": MetricName("my_metric"),
+                    },
+                },
+                {
+                    "~.*expr": {
+                        "name": MetricName("my_metric"),
+                    },
+                },
+            ],
+            {
+                MetricName("my_metric"),
+                MetricName("old_name_1"),
+            },
+            id="regex translation",
+        ),
+    ],
+)
+def test_reverse_translate_into_all_potentially_relevant_metrics(
+    canonical_name: MetricName,
+    current_version: int,
+    all_translations: Iterable[Mapping[MetricName, CheckMetricEntry]],
+    expected_result: frozenset[MetricName],
+) -> None:
+    assert (
+        _reverse_translate_into_all_potentially_relevant_metrics(
+            canonical_name,
+            current_version,
+            all_translations,
+        )
+        == expected_result
     )
